+++ sql/pg_extension_update_scripts_sequence.sql
@@ -2,6 +2,7 @@
 \o /dev/null
 \getenv extension_name EXTENSION_NAME
 \getenv extension_oldest_version EXTENSION_OLDEST_VERSION
+\getenv extension_version_v_suffix EXTENSION_VERSION_V_SUFFIX
 select
     not :{?extension_name} as extension_name_missing
     ,not :{?extension_oldest_version} as extension_oldest_version_missing
@@ -34,7 +35,17 @@ with version_path as (
 )
 ,update_script_path as (
     select
-        concat('sql/', :'extension_name', '--', v1.version, '--', v2.version, '.sql') as filename
+        concat(
+            'sql/', :'extension_name', '--', v1.version, '--', v2.version, '.sql'
+            ,case
+                when
+                    :'extension_version_v_suffix' ~ ('^:' || replace(v2.version, '.', '\.') || '\M')
+                then
+                    :'extension_version_v_suffix'
+                else
+                    ''
+            end
+        ) as filename_with_optional_suffix
         ,v1.ordinality  -- Doesn't matter if we pick `v1.version` or `v2.version`.
     from
         version_path as v1
@@ -43,7 +54,7 @@ with version_path as (
         on v2.ordinality = v1.ordinality + 1
 )
 select
-    string_agg(p.filename, ' ' order by p.ordinality) as file_paths
+    string_agg(p.filename_with_optional_suffix, ' ' order by p.ordinality) as file_paths
 from
     update_script_path as p
 \gset