+++ sql/pg_role_fkey_trigger_functions--1.0.0--1.0.1.sql
@@ -2,5 +2,7 @@
  * CHANGELOG.md:
  *
  * - This bugfix release only fixes the `CHANGELOG.md`, because, actually,
- *   version 1.0.0 _was_ release when version 1.0.0 was released.
+ *   version 1.0.0 _was_ released when version 1.0.0 was released.  [Except
+ *   that then this release repeated the exact same mistake for the release
+ *   that of _this_ version. 🤦]
  */